requirements.txt
streamlit==1.40.0
python-dotenv==1.0.1
psycopg[binary]==3.2.3
pgvector==0.3.6
sentence-transformers==3.2.1
groq==0.13.0
numpy==2.1.3
pandas==2.2.3
